Annotations added to User.java
